requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asyncio==4.0.0
certifi==2025.8.3
charset-normalizer==3.4.3
dedalus_labs==0.1.0a8
distro==1.9.0
elevenlabs==2.16.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
requests==2.32.5
sniffio==1.3.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
